+++ CMakeLists.txt
@@ -62,7 +62,6 @@ project(${BOARD_LIB_NAME}
 
 add_library(${PROJECT_NAME} STATIC
         src/HIB_Soon.cpp
-        targets/REV3-HIB/main_Soon.cpp
         include/HIB_Soon.h)
 
 # Add sources
